fix type of filesize output parameter of got_object_blob_dump_to_file()

+++ got/got.c
@@ -4438,7 +4438,7 @@ cmd_blame(int argc, char *argv[])
 	char *commit_id_str = NULL;
 	struct blame_cb_args bca;
 	int ch, obj_type, i;
-	size_t filesize;
+	off_t filesize;
 
 	memset(&bca, 0, sizeof(bca));
